clean up tests removing dependencies. Use constant instead of hardcoded string. testExportDropdown added
