+++ test/e2e/utils/page-utils.ts
@@ -1,7 +1,9 @@
-import { Page, Locator, expect } from '@playwright/test';
+import { expect, Locator, Page, test } from '@playwright/test';
 import fs from 'fs';
 
 // Locators
+import ExportLocators from '@isrd-isi-edu/chaise/test/e2e/locators/export';
+import ModalLocators from '@isrd-isi-edu/chaise/test/e2e/locators/modal';
 import PageLocators from '@isrd-isi-edu/chaise/test/e2e/locators/page';
 import RecordLocators from '@isrd-isi-edu/chaise/test/e2e/locators/record';
 import RecordsetLocators from '@isrd-isi-edu/chaise/test/e2e/locators/recordset';
@@ -140,10 +142,82 @@ export async function testButtonState(button: Locator, useSoftExpect: boolean, i
 
 export async function deleteDownloadedFiles(filePaths: string[]) {
   filePaths.forEach((path: string) => {
-      if (fs.existsSync(path)) {
-          // delete if there is any existing file with same name
-          fs.unlinkSync(path);
-          console.log(`file: ${path} has been removed`);
-      }
+    if (fs.existsSync(path)) {
+      // delete if there is any existing file with same name
+      fs.unlinkSync(path);
+      console.log(`file: ${path} has been removed`);
+    }
   });
 };
+
+/**
+ * function to test the export dropdown on record and recordset pages. If filenames has 3 values in it, then the
+ * 3rd test below will run checking the "configurations" submenu option
+ *
+ * @param fileNames string names for files to verify have downloaded
+ */
+export async function testExportDropdown(page: Page, fileNames: string[]) {
+  await test.step(`should have ${fileNames.length} options in the export dropdown menu.`, async () => {
+    const exportButton = ExportLocators.getExportDropdown(page);
+
+    await exportButton.click();
+    await expect.soft(ExportLocators.getExportOptions(page)).toHaveCount(fileNames.length);
+    // close the dropdown
+    await exportButton.click();
+  });
+
+  await test.step('should have "This record (CSV)" as a download option and download the file.', async () => {
+    await ExportLocators.getExportDropdown(page).click();
+
+    const csvOption = ExportLocators.getExportOption(page, 'This record (CSV)');
+    await expect.soft(csvOption).toHaveText('This record (CSV)');
+
+    await clickAndVerifyDownload(csvOption, fileNames[0]);
+  });
+
+  await test.step('should have "BDBag" as a download option and download the file.', async () => {
+    test.skip(!!process.env.CI, 'in CI the export server component is not configured and cannot be tested');
+
+    await ExportLocators.getExportDropdown(page).click();
+
+    const bagOption = ExportLocators.getExportOption(page, 'BDBag');
+    await expect.soft(bagOption).toHaveText('BDBag');
+
+    await clickAndVerifyDownload(bagOption, fileNames[1], async () => {
+      const modal = ModalLocators.getExportModal(page)
+      await expect.soft(modal).toBeVisible();
+      await expect.soft(modal).not.toBeAttached();
+    });
+  });
+
+  // NOTE: this is very specific to the test done in record/presentation.spec.ts
+  if (fileNames.length > 2) {
+    await test.step('should have "Configurations" option that opens a submenu to download the config file.', async () => {
+      test.skip(!!process.env.CI, 'in CI the export server component is not configured and cannot be tested');
+
+      await ExportLocators.getExportDropdown(page).click();
+
+      const configOption = ExportLocators.getExportOption(page, 'configurations');
+      await expect.soft(configOption).toHaveText('Configurations');
+
+      await configOption.click();
+
+      await expect.soft(ExportLocators.getExportSubmenuOptions(page)).toHaveCount(1);
+
+      const bdBagSubmenu = ExportLocators.getExportSubmenuOption(page, 'BDBag');
+      await expect.soft(bdBagSubmenu).toBeVisible();
+
+      await clickAndVerifyDownload(bdBagSubmenu, fileNames[2]); // use the last filename since it is the suggested filename
+
+      /**
+       * hover over to make the dropdown menu tooltip OverlayTrigger trigger so it will hide when another tooltip is shown in a later test
+
+       *
+       * NOTE: this is only an issue when `NODE_ENV="development"` since we are adding "focus" event for tooltips
+       *   this has no harm if the tooltip is not showing (node environment is production)
+       *   see /src/components/tooltip.tsx for more info
+       */
+      await ExportLocators.getExportDropdown(page).hover();
+    });
+  }
+}
